adds

+++ AtomicCore/Infrastructure/General/UUIDHandler.cs
@@ -16,7 +16,7 @@ namespace AtomicCore
             long i = 1;
 
             foreach (byte b in Guid.NewGuid().ToByteArray())
-                i *= ((int)b + 1);
+                i *= b + 1;
 
             return string.Format("{0:x}", i - DateTime.Now.Ticks);
         }
